requirements.txt
addict==2.4.0
numpy==1.21.6
boto3==1.24.48
botocore==1.27.48
certifi
charset-normalizer==2.1.0
cycler==0.11.0
Cython==0.29.32
Deprecated==1.2.13
filelock==3.8.0
fonttools==4.34.4
future==0.18.2
idna==3.3
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.5.2
MegEngine==1.10.0
mprop==0.16.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyarrow==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
smart-open==6.0.0
tabulate==0.8.10
terminaltables==3.1.10
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cu101
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
yapf==0.32.0
yacs==0.1.8
sklearn
# mmpycocotools==12.0.3
